refactor: replace clock package usage with DateTime.now() and remove unused clock dependency
